+++ app/src/main/java/com/company/mobile/android/appname/app/common/view/ClickActionFactory.kt
@@ -1,0 +1,8 @@
+package com.company.mobile.android.appname.app.common.view
+
+import android.view.View
+
+interface ClickActionFactory {
+
+    fun onClickAction(widget: View)
+}