+++ data/src/main/java/com/company/mobile/android/appname/data/bufferoo/repository/BufferooDataRepository.kt
@@ -1,8 +1,11 @@
 package com.company.mobile.android.appname.data.bufferoo.repository
 
-import com.company.mobile.android.appname.model.bufferoo.Bufferoo
 import com.company.mobile.android.appname.data.bufferoo.source.BufferooDataStoreFactory
 import com.company.mobile.android.appname.domain.bufferoo.repository.BufferooRepository
+import com.company.mobile.android.appname.model.bufferoo.Bufferoo
+import com.company.mobile.android.appname.model.bufferoo.Credentials
+import com.company.mobile.android.appname.model.bufferoo.SignedInBufferoo
+import com.company.mobile.android.appname.model.bufferoo.SignedOutBufferoo
 import io.reactivex.Completable
 import io.reactivex.Single
 
@@ -12,12 +15,12 @@ import io.reactivex.Single
  */
 class BufferooDataRepository(private val factory: BufferooDataStoreFactory) : BufferooRepository {
 
-    override fun clearBufferoos(): Completable {
-        return factory.retrieveCacheDataStore().clearBufferoos()
+    override fun signIn(username: String, password: String): Single<SignedInBufferoo> {
+        return factory.retrieveRemoteDataStore().signIn(username, password)
     }
 
-    override fun saveBufferoos(bufferoos: List<Bufferoo>): Completable {
-        return factory.retrieveCacheDataStore().saveBufferoos(bufferoos)
+    override fun getCredentials(): Single<Credentials> {
+        return factory.retrieveRemoteDataStore().getCredentials()
     }
 
     override fun getBufferoos(): Single<List<Bufferoo>> {
@@ -40,5 +43,17 @@ class BufferooDataRepository(private val factory: BufferooDataStoreFactory) : Bu
 
                 bufferooListSource
             }
+    }
+
+    override fun saveBufferoos(bufferoos: List<Bufferoo>): Completable {
+        return factory.retrieveCacheDataStore().saveBufferoos(bufferoos)
+    }
+
+    override fun clearBufferoos(): Completable {
+        return factory.retrieveCacheDataStore().clearBufferoos()
+    }
+
+    override fun signOut(): Single<SignedOutBufferoo> {
+        return factory.retrieveRemoteDataStore().signOut()
     }
 }