+++ data/src/main/java/com/company/mobile/android/appname/data/bufferoo/source/BufferooDataStore.kt
@@ -1,6 +1,9 @@
 package com.company.mobile.android.appname.data.bufferoo.source
 
 import com.company.mobile.android.appname.model.bufferoo.Bufferoo
+import com.company.mobile.android.appname.model.bufferoo.Credentials
+import com.company.mobile.android.appname.model.bufferoo.SignedInBufferoo
+import com.company.mobile.android.appname.model.bufferoo.SignedOutBufferoo
 import io.reactivex.Completable
 import io.reactivex.Single
 
@@ -11,13 +14,19 @@ import io.reactivex.Single
  */
 interface BufferooDataStore {
 
-    fun clearBufferoos(): Completable
+    fun signIn(username: String, password: String): Single<SignedInBufferoo>
+
+    fun getCredentials(): Single<Credentials>
+
+    fun getBufferoos(): Single<List<Bufferoo>>
 
     fun saveBufferoos(bufferoos: List<Bufferoo>): Completable
 
-    fun getBufferoos(): Single<List<Bufferoo>>
+    fun clearBufferoos(): Completable
 
     fun isValidCache(): Single<Boolean>
 
     fun setLastCacheTime(lastCache: Long)
+
+    fun signOut(): Single<SignedOutBufferoo>
 }