+++ datasources/src/test/java/com/company/mobile/android/appname/datasources/bufferoo/remote/BufferooRemoteImplTest.kt
@@ -1,6 +1,7 @@
 package com.company.mobile.android.appname.datasources.bufferoo.remote
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
+import androidx.test.platform.app.InstrumentationRegistry
 import com.company.mobile.android.appname.model.bufferoo.Bufferoo
 import com.company.mobile.android.appname.datasources.bufferoo.remote.mapper.BufferooEntityMapper
 import com.company.mobile.android.appname.datasources.bufferoo.remote.test.factory.BufferooFactory
@@ -16,19 +17,19 @@ class BufferooRemoteImplTest {
     private val entityMapper = mock<BufferooEntityMapper>()
     private val bufferooService = mock<BufferooService>()
 
-    private val bufferooRemoteImpl = BufferooRemoteImpl(bufferooService, entityMapper)
+    private val bufferooRemoteImpl = BufferooRemoteImpl(bufferooService, entityMapper, InstrumentationRegistry.getInstrumentation().targetContext)
 
     //<editor-fold desc="Get Bufferoos">
     @Test
     fun getBufferoosCompletes() {
-        stubBufferooServiceGetBufferoos(Single.just(BufferooFactory.makeBufferooResponse()))
+        stubBufferooServiceGetBufferoos(Single.just(BufferooFactory.makeBufferoosResponse()))
         val testObserver = bufferooRemoteImpl.getBufferoos().test()
         testObserver.assertComplete()
     }
 
     @Test
     fun getBufferoosReturnsData() {
-        val bufferooResponse = BufferooFactory.makeBufferooResponse()
+        val bufferooResponse = BufferooFactory.makeBufferoosResponse()
         stubBufferooServiceGetBufferoos(Single.just(bufferooResponse))
         val bufferooEntities = mutableListOf<Bufferoo>()
         bufferooResponse.team.forEach {
@@ -42,7 +43,7 @@ class BufferooRemoteImplTest {
 
     private fun stubBufferooServiceGetBufferoos(
         observable:
-        Single<BufferooService.BufferooResponse>
+        Single<BufferooService.BufferoosResponse>
     ) {
         whenever(bufferooService.getBufferoos())
             .thenReturn(observable)