+++ model/src/main/java/com/company/mobile/android/appname/model/bufferoo/Credentials.kt
@@ -1,0 +1,6 @@
+package com.company.mobile.android.appname.model.bufferoo
+
+/**
+ * Representation for a [Credentials] fetched from an external layer data source
+ */
+data class Credentials(val username: String, val id: String)